fix: handle big files

+++ includes/ft_ssl.h
@@ -6,7 +6,7 @@
 /*   By: ljoly <ljoly@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/04 18:03:23 by ljoly             #+#    #+#             */
-/*   Updated: 2018/10/11 17:44:21 by ljoly            ###   ########.fr       */
+/*   Updated: 2018/10/15 14:45:45 by ljoly            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,8 +28,8 @@ typedef struct	s_env
 {
 	char		*input;
 	size_t		input_len;
-	uint32_t	padding_bitsize;
-	uint32_t	input_bitsize;
+	size_t		padding_bitsize;
+	size_t		input_bitsize;
 	size_t		blocks;
 	uint32_t	*meta_block;
 	char		*output;
